+++ packages/ragbits-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Rename DocumentMeta create_text_document_from_literal to from_literal (#561)
 - Adjust typing for DocumentSearch (#554)
 - Add Prometheus & Grafana Monitoring for LLMs Using OpenTelemetry (#427)
 - Restructure audit module (#427)